Add workaround for Sony Ericsson Xperia Z3 Compact.
The `bbqscreen` USB service executable bundled with the app was
linked against a libstagefright.so where the AudioSource
constructor has the signature

        (audio_source_t, unsigned int, unsigned int)

while the library shipped on the Z3 Compact has

        (audio_source_t, unsigned int, unsigned int, int)

resulting in an unresolved symbol. We work around this issue by
injecting a shim library using LD_PRELOAD which intercepts the
constructor call and then calls the constructor in the shipped
library with a hardcoded 0 as the argument for the last extra
`int` parameter, which seems to work OK.

This is possible true for other models in the Xperia line as well.